+++ trunk/source/CACSL2BoogieTranslator/src/de/uni_freiburg/informatik/ultimate/cdt/translation/implementation/base/standardfunctions/StandardFunctionHandler.java
@@ -84,50 +84,50 @@ public class StandardFunctionHandler {
 			final ExpressionResultTransformer expressionResultTransformer, final LocationFactory locationFactory,
 			final ITypeHandler typeHandler, final CExpressionTranslator cEpressionTranslator,
 			final DataRaceChecker dataRaceChecker) {
-		final List<StandardFunctionHandler2> functionHandlers = List.of(
-				new LinuxStandardFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+		final List<FunctionModelProvider> functionHandlers = List.of(
+				new LinuxFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
 						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, settings,
 						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new PthreadStandardFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+				new PthreadFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
 						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, symboltable, settings,
 						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new StdioStandardFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+				new StdioFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
 						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, settings,
 						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new FloatMathStandardFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler,
+				new FloatMathFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, settings,
+						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
+				new GccBuiltinFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler,
 						expressionTranslation, memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes,
 						settings, expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new GccBuiltinStandardFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler,
-						expressionTranslation, memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes,
-						settings, expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new SvcompFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+				new SvcompFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
 						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, settings,
 						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new AtomicStandardfunctionHandler(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+				new AtomicFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
 						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, settings,
 						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new StringStandardFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+				new StringFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
 						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, settings,
 						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new StdlibStandardFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+				new StdlibFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
 						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, settings,
 						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new VariadicStandardFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler,
-						expressionTranslation, memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes,
-						settings, expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new AssertStandardFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+				new VariadicFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
 						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, settings,
 						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new TimeStandardFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+				new AssertFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
 						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, settings,
 						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new RandomStandardFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+				new TimeFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
 						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, settings,
 						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new SocketStandardFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+				new RandomFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
 						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, settings,
 						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
-				new SetjmpStandardFunctionHandler(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+				new SocketFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
+						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, settings,
+						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker),
+				new SetjmpFunctionModelProvider(functionTable, auxVarInfoBuilder, nameHandler, expressionTranslation,
 						memoryHandler, typeSizeAndOffsetComputer, procedureManager, typeSizes, settings,
 						expressionResultTransformer, typeHandler, cEpressionTranslator, dataRaceChecker));
 		mLogger = logger;
@@ -175,7 +175,7 @@ public class StandardFunctionHandler {
 	}
 
 	private static Map<String, IFunctionModelHandler>
-			getFunctionModels(final List<StandardFunctionHandler2> functionHandlers) {
+			getFunctionModels(final List<FunctionModelProvider> functionHandlers) {
 		final IFunctionModelHandler die = (main, node, loc, name) -> {
 			throw new UnsupportedSyntaxException(loc, "Unsupported function: " + name);
 		};
